Add settings panel for configuration, allow toggle for data publishing
