+++ src/assets/utils.scss
@@ -27,13 +27,13 @@
  * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
  * POSSIBILITY OF SUCH DAMAGE.
  */
-@media screen and (max-width: 1710px) {
+@media screen and (max-width: 1310px) {
   .hide-button-text {
     display: none;
   }
 }
 
-@media screen and (max-width: 1490px) {
+@media screen and (max-width: 1050px) {
   .hide-button-text-control {
     display: none;
   }
